+++ Codigo Fuente/src/main/java/es/uca/iw/ebz/Movimiento/Recibo/Recibo.java
@@ -5,6 +5,7 @@ import es.uca.iw.ebz.Movimiento.Movimiento;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Entity
@@ -17,7 +18,7 @@ public class Recibo {
 
     @Column(name = "Importe")
     @NotNull
-    private float fImporte;
+    private BigDecimal fImporte;
 
     @ManyToOne
     private Cuenta cuenta;
@@ -29,13 +30,13 @@ public class Recibo {
 
     //getters
     public UUID getId() {return id;}
-    public float getImporte() {return fImporte;}
+    public float getImporte() {return fImporte.floatValue();}
     public Cuenta getCuenta() {return cuenta;}
     public Movimiento getMovimiento() {return movimiento;}
 
     //setters
     public void setId(UUID id) {this.id = id;}
-    public void setImporte(float fImporte) {this.fImporte = fImporte;}
+    public void setImporte(float fImporte) {this.fImporte = BigDecimal.valueOf(fImporte);}
     public void setCuenta(Cuenta cuenta) {this.cuenta = cuenta;}
     public void setMovimiento(Movimiento movimiento) {this.movimiento = movimiento;}
 }
